+++ src/main/java/yilia/Type.java
@@ -1,0 +1,6 @@
+package yilia;
+public enum Type {
+    TODO,
+    DEADLINE,
+    EVENT
+}
