+++ src/main/java/yilia/command/MarkCommand.java
@@ -1,0 +1,23 @@
+package yilia.command;
+
+import yilia.Storage;
+import yilia.Ui;
+import yilia.task.TaskList;
+
+public class MarkCommand extends Command {
+    private final int index;
+
+    public MarkCommand(int index) {
+        this.index = index;
+    }
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        try {
+            tasks.get(index).check();
+            ui.showMarkStatus(tasks.get(index));
+        } catch (IndexOutOfBoundsException e) {
+            System.out.println("Index " + index + " out of bounds\nPlease input another index");
+        }
+    }
+}
