+++ src/main/java/yilia/command/UnmarkCommand.java
@@ -1,0 +1,23 @@
+package yilia.command;
+
+import yilia.Storage;
+import yilia.Ui;
+import yilia.task.TaskList;
+
+public class UnmarkCommand extends Command {
+    private final int index;
+
+    public UnmarkCommand(int index) {
+        this.index = index;
+    }
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        try {
+            tasks.get(index).uncheck();
+            ui.showUnmarkStatus(tasks.get(index));
+        } catch (IndexOutOfBoundsException e) {
+            System.out.println("Index " + index + " out of bounds\nPlease input another index");
+        }
+    }
+}
