+++ src/main/java/yilia/exception/TimeFormatException.java
@@ -1,0 +1,8 @@
+package yilia.exception;
+
+public class TimeFormatException extends Exception{
+    @Override
+    public String getMessage() {
+        return "Please input a date in the correct format";
+    }
+}
