+++ src/main/java/yilia/exception/YiliaException.java
@@ -1,0 +1,7 @@
+package yilia.exception;
+public class YiliaException extends Exception {
+    @Override
+    public String getMessage() {
+        return "☹ OOPS!!! I'm sorry, but I don't know what that means :-(";
+    }
+}
