+++ src/main/java/yilia/task/Todo.java
@@ -1,0 +1,17 @@
+package yilia.task;
+public class Todo extends Task {
+    public Todo(String content) {
+        super(content);
+    }
+    public Todo(String content, boolean isDone) {
+        super(content, isDone);
+    }
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+    @Override
+    public String parse() {
+        return "T / " + (isDone ? "1" : "0") + " / " + super.parse();
+    }
+}
